+++ src/i18n.js
@@ -152,7 +152,10 @@ function poToJson() {
           const addonItems = Pofile.parse(fs.readFileSync(addonlocale, 'utf8'))
             .items;
           items = [...addonItems, ...items];
-          console.log(`Merging ${addon} locales for ${lang}`);
+          if (require.main === module) {
+            // We only log it if called as script
+            console.log(`Merging ${addon} locales for ${lang}`);
+          }
         }
       });
     }